refactor(monocly): align browser state with observation
